Windows fixes (#138)
* * Working on fixing code portability issues for Visual Studio
* Fixing capnp build-related files for Windows

* * Fixed an error in Double with unknown implicit conversion

* * Updated gitignore to not include the .user file

* * Accidentally commented out sections to debug compile errors in VS
* Updated spacing in using_capnp.mpb

* * Updating using_capnp.mpb to use the git repo instead of the zip

* * Moved whatever is going on in the using_boost file (probably for
Android into the non-Windows section)

* * Removing a warning regarding type conversion

* * Fixed some warnings when compiling all tests in Windows
* Fixed errors related to CircularBuffer usage in Windows

+++ include/madara/knowledge/KnowledgeRecord.inl
@@ -806,7 +806,7 @@ inline uint32_t KnowledgeRecord::size(void) const
   {
     if (any_value_->supports_size())
     {
-      return any_value_->size();
+      return (uint32_t)any_value_->size();
     }
   }
   else if (type_ == BUFFER)
